Refactor tracking to middleware
